% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_swimlane.R
\name{geom_swimlane}
\alias{geom_swimlane}
\title{Swimlane charts}
\usage{
geom_swimlane(id_var, duration_var, cohort_var = NULL, ...)
}
\arguments{
\item{id_var}{Column name for subject identifier}

\item{duration_var}{Column name for numeric length of time on study}

\item{cohort_var}{Optional column name to sort by cohort}

\item{...}{Other options passed to geom_bar}
}
\value{
A ggplot2 compatible geom
}
\description{
A geom for swimlane plots
}
\examples{
library(ggplot2)

patient_disposition \%>\%
  order_swimlane(subject, weeks_on_study, cohort) \%>\%
  ggplot() +
  geom_swimlane(subject, weeks_on_study, cohort) +
  theme_swimlane(legend.position = c(.9, .1)) +
  ylab("Weeks on study")
}
