% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_swimlane_rug.R
\name{geom_swimlane_rug}
\alias{geom_swimlane_rug}
\title{Add a rug to the right-hand margin of a swimlane}
\usage{
geom_swimlane_rug(x, hjust = 1, label_var, labels = NULL, color = "black", ...)
}
\arguments{
\item{x}{Typically a subject identifier}

\item{hjust}{Used to adjust position in the margin}

\item{label_var}{Variable with the desired symbol or text annotation}

\item{labels}{A named vector transforming label_var to symbols (currently not used)}

\item{color}{Label colors (will accept a vector when the labels arg is activated)}

\item{...}{Further arguments to ggtext::geom_richtext()}
}
\description{
Add a rug to the right-hand margin of a swimlane
}
\examples{
library(ggplot2)

patient_disposition \%>\%
  dplyr::mutate(
    reason_off_study = tidyr::replace_na(reason_off_study, "&#8594;"),
    prior_drug = dplyr::case_when(
      prior_drug == "Yes"~ "&#9658;",
      TRUE ~ NA_character_
    )
  ) \%>\%
  order_swimlane(subject, weeks_on_study, cohort) \%>\%
  ggplot() +
  geom_swimlane(subject, weeks_on_study, cohort) +
  ggsci::scale_fill_jco() +
  geom_swimlane_text(subject, weeks_on_study, reason_off_study) +
  geom_point(
    aes(x = subject, y = partial_response),
    na.rm = TRUE
  ) +
  geom_swimlane_rug(x = subject, label_var = prior_drug, color = "#CD534C") +
  theme_swimlane(legend.position = c(.8, .1)) +
  ggtitle("Time on study by subject") +
  ylab("Weeks on study")
}
