% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_swimlane.R
\name{order_swimlane}
\alias{order_swimlane}
\title{Order a tibble for swimlane plotting}
\usage{
order_swimlane(.data, id_var, order_by, order_within = NULL)
}
\arguments{
\item{.data}{A tibble or data.frame}

\item{id_var}{Row (often subject) ID}

\item{order_by}{Variable to sort by, often time on study}

\item{order_within}{Variable to sort within, often cohort}
}
\value{
A tibble with \code{id_var} as a releveled factor
}
\description{
Assigns factor levels to an identifying key in a tibble (often subject ID)
by a second variable (often study duration), with the option to perform
this sorting within groups. Most often used before a call to ggplot() +
geom_swimlane()
}
\examples{
patient_disposition \%>\%
  order_swimlane(subject, weeks_on_study, cohort)
}
