% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_swimlane.R
\name{theme_swimlane}
\alias{theme_swimlane}
\title{A ggplot theme for swimlane plots}
\usage{
theme_swimlane(extra_margin_r = 15, ...)
}
\arguments{
\item{extra_margin_r}{Number (of pt units) to add to right margin, most
often to account for text annotations}

\item{...}{Further arguments to ggplot2::theme()}
}
\description{
A ggplot theme for swimlane plots
}
\examples{
library(ggplot2)

patient_disposition \%>\%
  order_swimlane(subject, weeks_on_study, cohort) \%>\%
  ggplot() +
  geom_swimlane(subject, weeks_on_study, cohort) +
  ggsci::scale_fill_jco() +
  theme_swimlane(legend.position = c(.8, .1))  \%+replace\%
  theme(axis.title.x = element_blank())
}
