% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_swimlane.R
\name{theme_swimlane}
\alias{theme_swimlane}
\title{A ggplot theme for swimlane plots}
\usage{
theme_swimlane(...)
}
\arguments{
\item{...}{Further arguments to ggplot2::theme()}
}
\description{
A ggplot theme for swimlane plots
}
\examples{
patient_disposition \%>\%
  order_swimlane(subject, weeks_on_study, cohort) \%>\%
  ggplot2::ggplot() +
  geom_swimlane(subject, weeks_on_study, cohort) +
  ggsci::scale_fill_jco() +
  theme_swimlane(legend.position = c(.8, .1)) +
  ggplot2::ggtitle("Time on study by subject") +
  ggplot2::ylab("\nWeeks on study")
}
